Linkify User Interactive Authentication errors (#12271)
