+++ src/components/views/dialogs/InteractiveAuthDialog.tsx
@@ -29,6 +29,7 @@ import InteractiveAuth, {
 } from "../../structures/InteractiveAuth";
 import { ContinueKind, SSOAuthEntry } from "../auth/InteractiveAuthEntryComponents";
 import BaseDialog from "./BaseDialog";
+import { Linkify } from "../../../Linkify";
 
 type DialogAesthetics = Partial<{
     [x in AuthType]: {
@@ -168,7 +169,9 @@ export default class InteractiveAuthDialog<T> extends React.Component<Interactiv
         if (this.state.authError) {
             content = (
                 <div id="mx_Dialog_content">
-                    <div role="alert">{this.state.authError.message || this.state.authError.toString()}</div>
+                    <Linkify>
+                        <div role="alert">{this.state.authError.message || this.state.authError.toString()}</div>
+                    </Linkify>
                     <br />
                     <AccessibleButton onClick={this.onDismissClick} className="mx_GeneralButton" autoFocus={true}>
                         {_t("action|dismiss")}
